renamed read_data.py to data_handler.py
